perf(cuvs): CPU scalar quantization for int8/uint8 IVF-PQ build
The int8/uint8 build path quantized one row at a time on the GPU
(cudaMalloc + H2D copy + transform kernel + D2H copy + handle.sync()
per row), crawling at ~1.1k rows/s on a 1M build. Scalar quantization
is a pure affine map from the trained [min,max], so once the quantizer
is trained no GPU is needed.

Add scalar_quantizer_t::transform_host<T>() -- a bit-for-bit CPU port of
cuVS' device quantize_op (double-domain scale/offset, lroundf ties, +128
for uint8) -- and use it in both quantize entry points:
  - add_chunk_float() trained path (f32 base): CPU-quantize straight into
    flattened_host_dataset, no per-row GPU round-trip.
  - flush_pending_float_chunks_internal() (f16 base / add_chunk_quantize):
    CPU transform per chunk instead of a per-chunk GPU transform.

Result on a 1M f32->uint8 build: ingest 880s -> 84s (~10.5x), total
CREATE INDEX 930s -> 105s (~8.8x), GPU idle during ingest, recall
preserved (0.838) and bytes identical to the device transform.

Also fix [B,Q]-refactor fallout in the test/bench suite:
  - pkg/cuvs/info_test.go: CAGRA/IVF-PQ int8/uint8 base int8 -> float32
    (the [B,Q] model has no 1-byte base; the quantizer cannot train on an
    int8 source).
  - benchmark_cuvs.cu / benchmark_filter.cu: gpu_*_t<T> -> <float,T>.
  - ivf_pq_test.cu f16 uint8-collapse repro test + standalone reproducers
    (uint8_quant_bug, wiki1m_uint8_bug) with Makefile targets.

+++ pkg/cuvs/info_test.go
@@ -149,7 +149,11 @@ func TestIndexInfoComprehensive(t *testing.T) {
 					bp := DefaultCagraBuildParams()
 					bp.IntermediateGraphDegree = 256
 					bp.GraphDegree = 128
-					index, err = NewGpuCagra[int8, int8](dataset, n_vectors, dimension, L2Expanded, bp, devices, 1, distMode, nil)
+					// int8 is a STORAGE (quantization) type, not a base type: the [B,Q]
+					// model only supports a float base (f32/f16) quantized to int8/uint8.
+					// Base f32, storage int8 (the wired f32xint8 combo); dataset is the
+					// storage type Q (see NewGpuCagra signature).
+					index, err = NewGpuCagra[float32, int8](dataset, n_vectors, dimension, L2Expanded, bp, devices, 1, distMode, nil)
 				case "IVF-Flat":
 					bp := DefaultIvfFlatBuildParams()
 					bp.NLists = 1000
@@ -158,7 +162,7 @@ func TestIndexInfoComprehensive(t *testing.T) {
 					bp := DefaultIvfPqBuildParams()
 					bp.NLists = 1000
 					bp.M = 16
-					index, err = NewGpuIvfPq[int8, int8](dataset, n_vectors, dimension, L2Expanded, bp, devices, 1, distMode, nil)
+					index, err = NewGpuIvfPq[float32, int8](dataset, n_vectors, dimension, L2Expanded, bp, devices, 1, distMode, nil)
 				}
 			case "uint8":
 				dataset := make([]uint8, n_vectors*uint64(dimension))
@@ -171,7 +175,8 @@ func TestIndexInfoComprehensive(t *testing.T) {
 					bp := DefaultCagraBuildParams()
 					bp.IntermediateGraphDegree = 256
 					bp.GraphDegree = 128
-					index, err = NewGpuCagra[uint8, uint8](dataset, n_vectors, dimension, L2Expanded, bp, devices, 1, distMode, nil)
+					// uint8 storage from a float base (wired f32xuint8 combo); see int8 note above.
+					index, err = NewGpuCagra[float32, uint8](dataset, n_vectors, dimension, L2Expanded, bp, devices, 1, distMode, nil)
 				case "IVF-Flat":
 					bp := DefaultIvfFlatBuildParams()
 					bp.NLists = 1000
@@ -180,7 +185,7 @@ func TestIndexInfoComprehensive(t *testing.T) {
 					bp := DefaultIvfPqBuildParams()
 					bp.NLists = 1000
 					bp.M = 16
-					index, err = NewGpuIvfPq[uint8, uint8](dataset, n_vectors, dimension, L2Expanded, bp, devices, 1, distMode, nil)
+					index, err = NewGpuIvfPq[float32, uint8](dataset, n_vectors, dimension, L2Expanded, bp, devices, 1, distMode, nil)
 				}
 			}
